+++ .github/workflows/ci.yml
@@ -76,7 +76,7 @@ jobs:
       - name: Install node
         uses: actions/setup-node@v1
         with:
-          node-version: '10.24.1'
+          node-version: '14.15.3'
       - uses: actions/setup-python@v2
         with:
           python-version: '2.7'
