+++ jest.config-standalone.js
@@ -1,3 +1,9 @@
 module.exports = {
+  preset: 'ts-jest',
+  globals: {
+    'ts-jest': {
+      tsconfig: 'tsconfig.json',
+    },
+  },
   roots: ['standalone-e2e'],
 }
